We have a layout!
There is a drop down menu now, and I also put in media queries under the style sheet with some basic changes to help demonstrate them.

However, it still needs a lot of extra polish - the media queries need better things to do than what they're doing, and we'll also have to experiment a little with the sizes they define...

I know the logo isn't perfect, but that's what I've got for now and that's better than what we had yesterday.